voice_assistant/requirements.txt
# ── Core ────────────────────────────────────────────
python-dotenv>=1.0,<2

# ── MQTT (필수) ─────────────────────────────────────
paho-mqtt>=1.6,<2

# ── Audio ───────────────────────────────────────────
sounddevice>=0.4.6,<1
webrtcvad-wheels>=2.0.10,<3

# ── TTS fallback ────────────────────────────────────
pyttsx3>=2.90,<3

# ── Wake-word (optional — porcupine 모드 전용) ──────
# pvporcupine>=3.0,<4
